several improvements to modules

+++ library/azure_rm_postgresqlserver.py
@@ -77,9 +77,6 @@ options:
     location:
         description:
             - The location the resource resides in.
-    tags:
-        description:
-            - Application-specific metadata in the form of key-value pairs.
 
 extends_documentation_fragment:
     - azure
@@ -109,7 +106,6 @@ EXAMPLES = '''
         admin_username: administrator_login
         admin_password: administrator_login_password
       location: location
-      tags: tags
 '''
 
 RETURN = '''
@@ -180,10 +176,6 @@ class AzureRMServers(AzureRMModuleBase):
                 type='str',
                 required=False
             ),
-            tags=dict(
-                type='dict',
-                required=False
-            ),
             state=dict(
                 type='str',
                 required=False,
@@ -211,18 +203,18 @@ class AzureRMServers(AzureRMModuleBase):
         for key in list(self.module_arg_spec.keys()) + ['tags']:
             if hasattr(self, key):
                 setattr(self, key, kwargs[key])
-            elif key == "sku" and kwargs[key] is not None:
-                self.parameters.update({"sku": kwargs[key]})
-            elif key == "properties" and kwargs[key] is not None:
-                self.parameters.update({"properties": kwargs[key]})
-            elif key == "location" and kwargs[key] is not None:
-                self.parameters.update({"location": kwargs[key]})
-            elif key == "tags" and kwargs[key] is not None:
-                self.parameters.update({"tags": kwargs[key]})
+            elif kwargs[key] is not None:
+                if key == "sku":
+                    self.parameters.update({"sku": kwargs[key]})
+                elif key == "properties":
+                    self.parameters.update({"properties": kwargs[key]})
+                elif key == "location":
+                    self.parameters.update({"location": kwargs[key]})
 
         self.adjust_parameters()
 
         old_response = None
+        response = None
         results = dict()
 
         self.mgmt_client = self.get_mgmt_svc_client(PostgreSQLManagementClient,
@@ -261,27 +253,25 @@ class AzureRMServers(AzureRMModuleBase):
                 self.results['changed'] = True
             else:
                 self.results['changed'] = old_response.__ne__(response)
-
-            self.results.update(response)
-
-            # remove unnecessary fields from return state
-            self.results.pop('name', None)
-            self.results.pop('type', None)
-            self.results.pop('location', None)
-            self.results.pop('tags', None)
-            self.results.pop('sku', None)
-            self.results.pop('administrator_login', None)
-            self.results.pop('storage_mb', None)
-            self.results.pop('ssl_enforcement', None)
             self.log("Creation / Update done")
         elif self.to_do == Actions.Delete:
             self.log("PostgreSQL Server instance deleted")
+
+            if self.check_mode:
+                return self.results
+
             self.delete_postgresqlserver()
             self.results['changed'] = True
         else:
             self.log("PostgreSQL Server instance unchanged")
-            self.results['state'] = old_response
             self.results['changed'] = False
+            response = old_response
+
+        if response is not None:
+            self.results["id"] = response["id"]
+            self.results["version"] = response["version"]
+            self.results["user_visible_state"] = response["user_visible_state"]
+            self.results["fully_qualified_domain_name"] = response["fully_qualified_domain_name"]
 
         return self.results
